Move http errors to http module, fix specs
