+++ lib/seam/auth.rb
@@ -2,9 +2,14 @@
 
 require_relative "options"
 require_relative "token"
-require_relative "errors"
 
 module SeamAuth
+  class SeamInvalidTokenError < StandardError
+    def initialize(message)
+      super("Seam received an invalid token: #{message}")
+    end
+  end
+
   def self.get_auth_headers(api_key: nil, personal_access_token: nil, workspace_id: nil)
     if SeamOptions.seam_http_options_with_api_key?(api_key: api_key, personal_access_token: personal_access_token)
       return get_auth_headers_for_api_key(api_key)
@@ -24,23 +29,23 @@ module SeamAuth
 
   def self.get_auth_headers_for_api_key(api_key)
     if SeamAuth.client_session_token?(api_key)
-      raise Http::SeamInvalidTokenError.new(
+      raise SeamInvalidTokenError.new(
         "A Client Session Token cannot be used as an api_key"
       )
     end
 
-    raise Http::SeamInvalidTokenError.new("A JWT cannot be used as an api_key") if SeamAuth.jwt?(api_key)
+    raise SeamInvalidTokenError.new("A JWT cannot be used as an api_key") if SeamAuth.jwt?(api_key)
 
-    raise Http::SeamInvalidTokenError.new("An Access Token cannot be used as an api_key") if SeamAuth.access_token?(api_key)
+    raise SeamInvalidTokenError.new("An Access Token cannot be used as an api_key") if SeamAuth.access_token?(api_key)
 
     if SeamAuth.publishable_key?(api_key)
-      raise Http::SeamInvalidTokenError.new(
+      raise SeamInvalidTokenError.new(
         "A Publishable Key cannot be used as an api_key"
       )
     end
 
     unless SeamAuth.seam_token?(api_key)
-      raise Http::SeamInvalidTokenError.new(
+      raise SeamInvalidTokenError.new(
         "Unknown or invalid api_key format, expected token to start with #{SeamAuth::TOKEN_PREFIX}"
       )
     end
@@ -50,25 +55,25 @@ module SeamAuth
 
   def self.get_auth_headers_for_personal_access_token(personal_access_token, workspace_id)
     if SeamAuth.jwt?(personal_access_token)
-      raise Http::SeamInvalidTokenError.new(
+      raise SeamInvalidTokenError.new(
         "A JWT cannot be used as a personal_access_token"
       )
     end
 
     if SeamAuth.client_session_token?(personal_access_token)
-      raise Http::SeamInvalidTokenError.new(
+      raise SeamInvalidTokenError.new(
         "A Client Session Token cannot be used as a personal_access_token"
       )
     end
 
     if SeamAuth.publishable_key?(personal_access_token)
-      raise Http::SeamInvalidTokenError.new(
+      raise SeamInvalidTokenError.new(
         "A Publishable Key cannot be used as a personal_access_token"
       )
     end
 
     unless SeamAuth.access_token?(personal_access_token)
-      raise Http::SeamInvalidTokenError.new(
+      raise SeamInvalidTokenError.new(
         "Unknown or invalid personal_access_token format, expected token to start with #{SeamAuth::ACCESS_TOKEN_PREFIX}"
       )
     end
@@ -81,25 +86,25 @@ module SeamAuth
 
   def self.get_auth_headers_for_multi_workspace_personal_access_token(personal_access_token)
     if SeamAuth.jwt?(personal_access_token)
-      raise Http::SeamInvalidTokenError.new(
+      raise SeamInvalidTokenError.new(
         "A JWT cannot be used as a personal_access_token"
       )
     end
 
     if SeamAuth.client_session_token?(personal_access_token)
-      raise Http::SeamInvalidTokenError.new(
+      raise SeamInvalidTokenError.new(
         "A Client Session Token cannot be used as a personal_access_token"
       )
     end
 
     if SeamAuth.publishable_key?(personal_access_token)
-      raise Http::SeamInvalidTokenError.new(
+      raise SeamInvalidTokenError.new(
         "A Publishable Key cannot be used as a personal_access_token"
       )
     end
 
     unless SeamAuth.access_token?(personal_access_token)
-      raise Http::SeamInvalidTokenError.new(
+      raise SeamInvalidTokenError.new(
         "Unknown or invalid personal_access_token format, expected token to start with #{SeamAuth::ACCESS_TOKEN_PREFIX}"
       )
     end
