+++ spec/seam_client/personal_access_token_spec.rb
@@ -7,19 +7,19 @@ RSpec.describe Seam::Client do
     it "raises error for invalid personal access token formats" do
       expect do
         Seam.from_personal_access_token("some-invalid-key-format", workspace_id)
-      end.to raise_error(Http::SeamInvalidTokenError, /Unknown/)
+      end.to raise_error(SeamAuth::SeamInvalidTokenError, /Unknown/)
 
       expect do
         Seam.from_personal_access_token("seam_apikey_token", workspace_id)
-      end.to raise_error(Http::SeamInvalidTokenError, /Unknown/)
+      end.to raise_error(SeamAuth::SeamInvalidTokenError, /Unknown/)
 
       expect do
         Seam.from_personal_access_token("seam_cst", workspace_id)
-      end.to raise_error(Http::SeamInvalidTokenError, /Client Session Token/)
+      end.to raise_error(SeamAuth::SeamInvalidTokenError, /Client Session Token/)
 
       expect do
         Seam.from_personal_access_token("ey", workspace_id)
-      end.to raise_error(Http::SeamInvalidTokenError, /JWT/)
+      end.to raise_error(SeamAuth::SeamInvalidTokenError, /JWT/)
     end
   end
 end
