+++ spec/spec_helper.rb
@@ -6,6 +6,7 @@ require "simplecov-console"
 SimpleCov.start
 
 require "seam"
+require "seam/http"
 require "seam/auth"
 require "seam/options"
 require "seam/client_multi_workspace"
